requirements.txt
fastapi
uvicorn
httpx
python-dotenv
requests
beautifulsoup4
lxml
openai